Optimize matrix multiplication using MKL by changing the data type from 'short' to 'float'

+++ cpp/src/include/matmul.h
@@ -12,12 +12,12 @@
 using namespace arma;
 using namespace std;
 
-Mat<short> parallel_matmul(const Mat<short> &A, const Mat<short> &B, int node, int n_procs);
-Mat<short> parallel_matsq(const Mat<short> &A, const Mat<short> &B, int node, int n_procs);
-void mat2array(short **A, const Mat<short> &B);
-void array2mat(short **A, Mat<short> &B);
-void mat2array_square(short **A, const Mat<short> &B);
-void array2mat_square(short **A, Mat<short> &B);
+Mat<float> parallel_matmul(const Mat<float> &A, const Mat<float> &B, int node, int n_procs);
+Mat<float> parallel_matsq(const Mat<float> &A, const Mat<float> &B, int node, int n_procs);
+void mat2array(float **A, const Mat<float> &B);
+void array2mat(float **A, Mat<float> &B);
+void mat2array_square(float **A, const Mat<float> &B);
+void array2mat_square(float **A, Mat<float> &B);
 void test_matmul(int rows, int mode = 0, string mat_A = "", string mat_B = "", string mat_C = "");
 void test_matsq(int rows, int mode = 0, string mat_A = "", string mat_C = "");
 void test_parallel_matmul(int rows, int *argc, char ***argv, int mode = 0, string mat_A = "", string mat_B = "", string mat_C = "");
